requirements.txt
hazelcast-python-client==5.1